Finished refactoring authentication/authorization (for now)
