+++ EmployeeReimbursementSystem/ApiLayer/Program.cs
@@ -21,6 +21,7 @@ public class Program
         builder.Services.AddScoped<ITicketService, TicketService>();
         builder.Services.AddScoped<IEmployeeRepository, EmployeeRepository>();
         builder.Services.AddScoped<ITicketRepository, TicketRepository>();
+        builder.Services.AddScoped<IAuthRepository, AuthRepository>();
         builder.Services.AddScoped<IEmployeeValidationService, EmployeeValidationService>();
         builder.Services.AddScoped<ITicketValidationService, TicketValidationService>();
         builder.Services.AddScoped<IEmployeeAuthService, EmployeeAuthService>();
